pyproject.toml
[tool.poetry]
name = "publication-analytics"
version = "0.1.0"
description = ""
authors = ["Adam J Hockenberry"]

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.4.2"
biopython = "^1.79"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
